refactor: move filters to seperate hook
