+++ packages/raystack/v1/components/data-table/components/filters.tsx
@@ -6,11 +6,8 @@ import { FilterIcon } from "../../icons";
 import { DropdownMenu } from "../../dropdown-menu";
 import { useDataTable } from "../hooks/useDataTable";
 import { Flex } from "../../flex";
-import {
-  FilterOperatorTypes,
-  FilterType,
-  filterOperators,
-} from "~/v1/types/filters";
+import { FilterOperatorTypes } from "~/v1/types/filters";
+import { useFilters } from "../hooks/useFilters";
 
 interface AddFilterProps<TData, TValue> {
   columnList: DataTableColumn<TData, TValue>[];
@@ -56,78 +53,15 @@ function AddFilter<TData, TValue>({
 }
 
 export function Filters<TData, TValue>() {
-  const { table, updateTableQuery, tableQuery } = useDataTable();
+  const { table, tableQuery } = useDataTable();
   const columns = table?.getAllColumns() as DataTableColumn<TData, TValue>[];
 
-  function onAddFilter(column: DataTableColumn<TData, TValue>) {
-    const columnDef = column.columnDef;
-    const id = columnDef.accessorKey || column.id;
-    const options = columnDef.filterOptions || [];
-    const columnType = columnDef.columnType || "text";
-    const defaultFilter = filterOperators[columnType][0];
-    const defaultValue =
-      columnType === FilterType.date
-        ? new Date()
-        : columnType === FilterType.select
-        ? options[0].value
-        : "";
-
-    updateTableQuery((query) => {
-      return {
-        ...query,
-        filters: [
-          ...(query.filters || []),
-          // TODO: Add default filter value in column definition
-          {
-            _type: columnType,
-            name: id,
-            value: defaultValue,
-            operator: defaultFilter.value,
-          },
-        ],
-      };
-    });
-  }
-
-  function handleRemoveFilter(columnId: string) {
-    updateTableQuery((query) => {
-      return {
-        ...query,
-        filters: query.filters?.filter((filter) => filter.name !== columnId),
-      };
-    });
-  }
-
-  function handleFilterValueChange(columnId: string, value: any) {
-    updateTableQuery((query) => {
-      return {
-        ...query,
-        filters: query.filters?.map((filter) => {
-          if (filter.name === columnId) {
-            return { ...filter, value };
-          }
-          return filter;
-        }),
-      };
-    });
-  }
-
-  function handleFilterOperationChange(
-    columnId: string,
-    operator: FilterOperatorTypes
-  ) {
-    updateTableQuery((query) => {
-      return {
-        ...query,
-        filters: query.filters?.map((filter) => {
-          if (filter.name === columnId) {
-            return { ...filter, operator };
-          }
-          return filter;
-        }),
-      };
-    });
-  }
+  const {
+    onAddFilter,
+    handleRemoveFilter,
+    handleFilterValueChange,
+    handleFilterOperationChange,
+  } = useFilters<TData, TValue>();
 
   const columnList = columns?.filter(
     (column) => column.columnDef.enableColumnFilter
